refactor(ui): break sidebar layout import cycle
